[moe] fix moe bugs (#1633)
